Remove components + Update Sidebar component
