+++ src/components/layout/sidebar/index.tsx
@@ -1,2 +1,1 @@
 export * from './sidebar'
-export * from './list'
