--- src/components/layout/sidebar/list/file/index.tsx
@@ -1,1 +1,0 @@
-export * from './file'
