--- src/components/layout/sidebar/list/files/files-styles.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components/macro'
-
-export const FilesWrapper = styled.nav``
-
-export const FilesList = styled.ul`
-  list-style: none;
-`
