--- src/components/layout/sidebar/list/files/files.tsx
@@ -1,24 +1,0 @@
-// Components
-import { File } from 'components/layout'
-
-// Styles
-import * as S from './files-styles'
-
-// Types & Props
-import { FileProps } from 'resources/types'
-
-type FilesProps = {
-  files: FileProps[]
-}
-
-const FileList = ({ files }: FilesProps) => (
-  <S.FilesWrapper>
-    <S.FilesList>
-      {files.map((file) => (
-        <File key={file.id} {...file} />
-      ))}
-    </S.FilesList>
-  </S.FilesWrapper>
-)
-
-export { FileList }
