--- src/components/layout/sidebar/list/files/index.tsx
@@ -1,1 +1,0 @@
-export * from './files'
