--- src/components/layout/sidebar/list/index.tsx
@@ -1,2 +1,0 @@
-export * from './file'
-export * from './files'
